Trying to use my shared memory hash table
